Keep page status on stats
